chore/feat: refactor CLI UX + hint about proc sync with dump #4454
