+++ lib/API.js
@@ -23,7 +23,7 @@ var KMDaemon    = require('@pm2/agent/src/InteractorClient');
 var Config      = require('./tools/Config');
 var Modularizer = require('./API/Modules/Modularizer.js');
 var path_structure = require('../paths.js');
-var UX          = require('./API/CliUx');
+var UX          = require('./API/UX');
 var pkg         = require('../package.json');
 var hf = require('./API/Modules/flagExt.js');
 var Configuration = require('./Configuration.js');
@@ -629,7 +629,7 @@ class API {
           process.stdout.write('\x1b[0f');
           console.log('Last refresh: ', moment().format('LTS'));
           that.Client.executeRemote('getMonitorData', {}, function(err, list) {
-            UX.dispAsTable(list, null);
+            UX.list(list, null);
           });
         }
 
@@ -1748,10 +1748,10 @@ class API {
         return that.exitCli(conf.ERROR_EXIT);
       }
       if (process.stdout.isTTY === false) {
-        UX.miniDisplay(list);
+        UX.list_min(list);
       }
       else if (commander.miniList && !commander.silent)
-        UX.miniDisplay(list);
+        UX.list_min(list);
       else if (!commander.silent) {
         if (that.gl_interact_infos) {
           var dashboard_url = `https://app.pm2.io/#/r/${that.gl_interact_infos.public_key}`
@@ -1765,7 +1765,7 @@ class API {
                           chalk.bold(that.gl_interact_infos.machine_name),
                           chalk.bold(dashboard_url))
         }
-        UX.dispAsTable(list, sys_infos);
+        UX.list(list, sys_infos);
         //Common.printOut(chalk.white.italic(' Use `pm2 show <id|name>` to get more details about an app'));
       }
 
@@ -1880,7 +1880,7 @@ class API {
 
       if (!cb) {
         found_proc.forEach(function(proc) {
-          UX.describeTable(proc);
+          UX.describe(proc);
         });
       }
