+++ lib/API/Configuration.js
@@ -1,7 +1,7 @@
 
 var Common               = require('../Common.js');
 var cst                  = require('../../constants.js');
-var UX                   = require('./CliUx');
+var UX                   = require('./UX');
 var chalk                = require('chalk');
 var Configuration        = require('../Configuration.js');
 
@@ -186,7 +186,7 @@ module.exports = function(CLI) {
 };
 
 function interactiveConfigEdit(cb) {
-  UX.openEditor(cst.PM2_MODULE_CONF_FILE, function(err, data) {
+  UX.helpers.openEditor(cst.PM2_MODULE_CONF_FILE, function(err, data) {
     Common.printOut(chalk.bold('Module configuration (%s) edited.'), cst.PM2_MODULE_CONF_FILE);
     Common.printOut(chalk.bold('To take changes into account, please restart module related.'), cst.PM2_MODULE_CONF_FILE);
     if (err)
@@ -206,7 +206,7 @@ function displayConf(target_app, cb) {
   }
 
   Configuration.getAll(function(err, data) {
-    UX.dispKeys(data, target_app);
+    UX.helpers.dispKeys(data, target_app);
     return cb();
   });
 }
