+++ lib/API/Extra.js
@@ -7,7 +7,7 @@
 
 var cst         = require('../../constants.js');
 var Common      = require('../Common.js');
-var UX          = require('./CliUx');
+var UX          = require('./UX');
 var chalk       = require('chalk');
 var path        = require('path');
 var fs          = require('fs');
@@ -134,7 +134,7 @@ module.exports = function(CLI) {
 
         fmt.sep();
         fmt.title(chalk.bold.blue('PM2 list'));
-        UX.dispAsTable(list, that.gl_interact_infos);
+        UX.list(list, that.gl_interact_infos);
 
         fmt.sep();
         fmt.title(chalk.bold.blue('Daemon logs'));
