+++ lib/API/Modules/index.js
@@ -7,7 +7,6 @@
 
 var cst          = require('../../../constants.js');
 var Common       = require('../../Common.js');
-var UX           = require('../CliUx');
 var chalk        = require('chalk');
 var forEachLimit = require('async/forEachLimit');
 
@@ -119,47 +118,3 @@ module.exports = function(CLI) {
     });
   };
 };
-
-// Special module with post display
-function postDisplay(app, cb) {
-  var that = this;
-  var retry = 0;
-
-  UX.processing.start('Initializing module');
-
-  (function detectModuleInit() {
-    retry++;
-    if (retry > 12) {
-      // Module init has timeouted
-      return displayOrNot(null);
-    }
-    that.describe(app.pm_id, function(err, data) {
-
-      if (data && data[0] && data[0].pm2_env &&
-          data[0].pm2_env.axm_options &&
-          data[0].pm2_env.axm_options.human_info) {
-        return displayOrNot(data[0]);
-      }
-      setTimeout(function() {
-        detectModuleInit();
-      }, 300);
-    });
-  })();
-
-  function displayOrNot(app) {
-    UX.processing.stop();
-
-    if (app) {
-      var module_name = app.name;
-      var human_info = app.pm2_env.axm_options.human_info;
-
-      UX.postModuleInfos(module_name, human_info);
-      Common.printOut(chalk.white.italic(' Use `pm2 show %s` to display this helper'), module_name);
-      Common.printOut(chalk.white.italic(' Use `pm2 logs %s [--lines 1000]` to display logs'), module_name);
-      Common.printOut(chalk.white.italic(' Use `pm2 monit` to monitor CPU and Memory usage'), module_name);
-      return cb ? cb(null, app) : that.exitCli(cst.SUCCESS_EXIT);
-    }
-
-    return cb ? cb(null, { msg : 'Module started' }) : that.speedList(cst.SUCCESS_EXIT);
-  }
-}
